+++ JobApplication/src/main/java/cr/ac/una/job/exceptions/EmpresaNotFoundException.java
@@ -1,7 +1,7 @@
 package cr.ac.una.job.exceptions;
 
 public class EmpresaNotFoundException extends RuntimeException {
-    public EmpresaNotFoundException(Integer id) {
+    public EmpresaNotFoundException(Long id) {
         super("Empresa con id " + id + " no encontrada");
     }
 }
