+++ JobApplication/src/main/java/cr/ac/una/job/exceptions/UsuarioNotFoundException.java
@@ -1,7 +1,7 @@
 package cr.ac.una.job.exceptions;
 
 public class UsuarioNotFoundException extends RuntimeException {
-    public UsuarioNotFoundException(Integer id) {
+    public UsuarioNotFoundException(Long id) {
         super("Usuario con id " + id + " no encontrado");
     }
 }
